Separate toolbar elements within separate modules
