+++ core/app/globals.css
@@ -282,6 +282,10 @@ h6.tiptap  {
   font-size: 1rem;
 }
 
+p.tiptap {
+  margin-bottom:1em;
+}
+
 table {
   border-collapse: collapse;
   margin: 0;
