--- core/components/learning/courses/course-editor/CourseEditor.tsx
@@ -1,1 +1,0 @@
-"use client";
